# [TensorIR] Support for L2 prefetch async copy and pred_guard enabled async in vectorized if_then_else