@Hzfengsy 2023-03-18
Thanks @LeiWang1999 for the improvement. Could you please add some performance benchmarks as it is a perf-related PR :)