@junrushao 2023-03-18
That makes a lot of sense. Thanks @LeiWang1999 for the bugfix! In addition, I am curious too if L2 prefetch could really help with performance :-) 

@cblmemo @andy-yang-1 PTAL