@andy-yang-1 2023-03-19
@LeiWang1999  Yeah, you are right. The ptx_ldg32 pass only supports ldg32 instruction for loading 4 bytes from global memory. I will also add support for them in the future :wink: I confused this with the automatic async work done by Tian. This work does not need to consider the architecture issue. Thank you very much for your feedback!