@LeiWang1999 2023-03-22
I got all test passed in my branch, however, after merge this pull request, which merged into main two days ago :
 https://github.com/apache/tvm/commit/c6c89c3a252b92c379d79b9b8fdbccf2339b131c

https://github.com/apache/tvm/blob/c6c89c3a252b92c379d79b9b8fdbccf2339b131c/src/tir/transforms/lower_async_dma.cc#L53-L57

add some constraints into the flow of dma lowering, it seems like this pr is made for Hexagon backend, but it will also effect on cuda codegen..

CC~ @junrushao @adstraw