@LeiWang1999 2023-04-02
I think this pr is ready to be merged in if I resolved the last lint ci failure (actually the link checker ci told me which file that do not follow the format rule, but I didn't find any scirpt that can auto-format the file, like clang-format or pep8, I may need some helps. @junrushao 

@adstraw I have a tricky fix of the dma lowering conflict between cuda backend and Hexagon, in the latest commit  `e08fff9aeca293325d1c81734dea2543ddd7ce26` and `` non contiguous memory access will not be a illegal behavior and it will fall back to a case that cuda required, now it seems that this chage works, but I think in the long terms view, we should separate the asynchronous copy of gpu from the dma of cpu