@junrushao 2023-04-02
@LeiWang1999 To run linters locally, below are some commands you may find useful:

```python
# Assume you are now at TVM's root directory

# Run all linters
bash ./tests/scripts/task_lint.sh
# Run python's formatter: black. CI uses black 22.12.0.
bash ./tests/lint/git-black.sh
# Run c++'s formatter: clang-format
bash ./tests/lint/git-clang-format.sh
# Run whitespace checking, another linter unhappy with your code
bash ./tests/lint/whitespace.sh
```

You might need to ensure the version of black/clang-format are consistent with our CI - which can be a bit annoying at times.